Bug fixed : Use Regular Expression to test for a decimal.
Previously, I checked if the last char of the value is "." then user could not input another ".". Ex: 10.. was not a possible screen value which was good.
Unfortunately, 10.0.  was possible because I exclusively checked the last char. So, I use Regular Expression to check if its contain decimal instead, in order to cover all cases.